Example of coding in R for research
This is an excerpt of code from the research project for computational linguistics that I work on. Variable names are changed, and the data being dealt with or other dependencies are not included to maintain overall privacy. The code will not run without all of that, however it is to provide an example of the kind of coding I do for this project. This file handles textual data. It creates complex networks with measures generated to analyse the connectedness of human-written pieces of text. The research is ongoing, and this is just one excerpt from a larger codebase.